[ADD] own module for jax kernels
